+++ inst/tmp/tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(rOpenserver)
-
-test_check("rOpenserver")